More kast <-> kore conversions (#83)
This PR makes a few changes:

- The default sort for symbols in `prelude.ml` is changed from `K` to
`GENERATED_TOP_CELL`, which is more frequently what we want.
- Conversions for ML symbols `#And` and `#Equals` are added, which
enables us to begin dealing with constrained terms passed from `CTerm`
class.
- When retrieving productions from the definition by `KLabel`, we only
compare the klabel name and not the sorts. This should still be safe,
because we're still asserting that only one production is returned. So
if the same symbol has different signatures, this method will die
horribly and the user can make it smarter or submit a bug report. So far
it hasn't caused me problems.

Co-authored-by: devops <devops@runtimeverification.com>