Add ability to to hide cells and node hashes in `KCFGShow.show()` (#282)
This PR adds some optional parameters to `KCFGShow.show()` which allow
the user to hide the node hashes for the nodes in `nodes`, as well as to
hide specified cells, which applies to both the printed nodes, node
deltas, and the generated module.

This is intended to be used in K foundry integration to have the option
to omit certain information from the `foundry-show` command output that
is likely to change for a test case without the behavior of that test
case changing. This would be used for certain tests to verify the KCFG
of that test does not change modulo node hashes and certain "volatile"
cells.

---------

Co-authored-by: devops <devops@runtimeverification.com>