Refactor KCFG printing, remove verified edge attribute (#302)
This PR is pulled out of:
https://github.com/runtimeverification/pyk/pull/124
In that PR, the `pretty_segments` routine is adjusted to accomodate the
new `KCFG.Split` edge. The refactoring of that routine can be done
standalone, by detecting splits in the old representation (which is what
this PR does):

This PR:

- Removes the `verified` attribute from `KCFG` datastructure, as it's
not longer as meaningful since we're using RPC based prover.
- Refactors the `_print_subgraph` method of `pretty_segments` to be a
bit more readable:
- Factors out `_print_split_edge` for printing out edge information,
which now is more informative.
- Prints out the current node _first_ (instead of at end of function,
then prints out successors).
  - Improves the formatting of the printed CFGs a bit.
- Factors out node sorter to decide which components of the KCFG to
print out first (focuses on printing out graphs first, then remaining
nodes second).
- Adds tests of KCFG printing cases that were not exercised before, and
broke during this refactoring (and were fixed), including:
  - Test of a `frontier` node being printed in graph.
  - Test of a `stuck` node being printed in graph.
  - Test of the user supplying the `node_printer` optional argument.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>