Refactors to KCFG edge representations (#404)
This PR is pulled out of:
https://github.com/runtimeverification/pyk/pull/329

The goal (and observation) here is that there is a lot of common
functionality between both normal edges and multi-edges. In particular,
both of them can define `targets` variable, which can simplify routines
which are doing traversals of the graph.

- Abstract method `KCFG.Successor.targets` is added, and implemented for
all successor types.
- Method `KCFG.Split.target_ids` is promoted to
`KCFG.Successor.target_ids`, so all classes have access to it.
- Method `KCFG.successors` now takes arguments indicating which types of
successors should be included in the list (default is all of them).
- Method `KCFG.predecessors` is added, which is the dual of
`KCFG.successors` and takes the same arguments.
- Method `KCFG.reachable_nodes` uses the new
`KCFG.{successors/predecessors}` to good effect to simplify quite a bit.
- Method `KCFG.Split.with_single_target` is added, which returns a new
`Split` with only the one target in place (used in `paths_between`).
- `KCFG.Split` now names its argument `_targets` instead, so that it can
match the signature of `@property targets`, but still exposes `@property
splits` as a dictionary.
- `KCFG.prune` now returns the list of pruned nodes, makes sure to
traverse covers as well.
- Makes `_print_split_edge` handle projecting out the components of the
`KCFG.Split` node for printing directly, so that the same interface
should work for printing `KCFG.NDBranch` with minimal modification.

---------

Co-authored-by: devops <devops@runtimeverification.com>