Implement outer parser (#550)
Implements outer syntax and parser.

Not accounting for tokenization, `domains.k` is parsed in ~0.015s. With
tokenization, it is parsed in ~0.06s.

Known differences to the K Frontend:
* `UserList` and `Lexical` are productions, not production items.
`UserList` is expected to be the only item in its production. `Lexical`
does not seem to work with other items in the production.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>