Factor out predicate to Subst logic into `Subst.from_pred` function (#573)
Closes #525 
Minor refactoring to #519.
As @palinatolmach pointed out, moving this logic into a static function
of `Subst` creates circular dependencies, specifically with the
`flatten_label` function. I'm not sure if it would be better to just
re-implement the function as I've done here or to have the function
`Subst.from_pred` as something like `subst_from_pred` in
`pyk/kast/manip.py`.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>