Refactor of the `KCFGViewer` (#578)
Going in a more step-by-step fashion compared to
https://github.com/runtimeverification/pyk/pull/460
Rolling out our own key bindings was messing up with the defaults of
textual and adding a lot of unecessary boilerpalte. This PR also closes
it.

- Dedicated classes for each widget in the `NodeView` (Term, Constraint,
Info, Custom)
- Relevant ones all inherits from `ScrollableContainer` and can be
focused. Adds native key bindings and can be selected using `Tab` or
`Tab+Shift` to loop the other way around
- using an absolute path for the css file so that the class
`APRProofViewer` inherits `KCFGViewer`'s css. `@import` is not something
that works with textual, but as we may want in the future to add stuff
only for the `APRProofViewer` css.

---------

Co-authored-by: François Guyot <francois.guyot@runtimeverification.com>
Co-authored-by: devops <devops@runtimeverification.com>