Handle empty response exception from the kore client (#595)
When the response from the `JsonRpcClient` is empty (usually happens
when there is a memory leak in the backend), no handling of the response
is done and the proof will crash on a JsonDecodeError that may not be
very meaningful for the user.
This PR is a fix for that that returns a `JsonRpcError` with the
appropriate code and tests.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>