Add ksequence under all K arguments under all productions (#639)
Closes #630
Issue originally discovered in #565 and fixed in #569, but this turned
out to be a special case of the more general case of kast-to-kore
transformations on productions with `K`-sort arguments. (`==K` is the
only built-in production like this, but the user can define their own.
This PR handles the more general case.)
Future work should be done via #653 to find and eliminate any other
potential problems with `kast_to_kore`.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>