Implement first past of KFlatModule conversion to Kore (#835)
This PR is to begin a refactoring needed for CSE, and also begin work on
implementing the skeleton of ModuleToKORE in pyk.

- A new routine is added to `konvert` for `kflatmodule_to_kore`, which
converts modules which contain only rules and imports to Kore.
- This routine is used for building up the dependencies and
circularities modules in initializing the `APRProver`, and then putting
them into the `add_module` endpoint directly.
- The code `KCFG.as_claim` is turned into a simpler `KCFG.as_rule`,
which uses existing routines instead of making a new one. This uses the
`cterm` modules `build_rule` functionality, which is adjusted to not
strip out information needed for making backend-compatible rules
(previously this routine was used for display purposes only).
- Old code for handling the dependencies/circularities modules in
`APRProver` is removed.

---------

Co-authored-by: devops <devops@runtimeverification.com>