Print parseable rules from KCFGShow (#857)
In this PR: https://github.com/runtimeverification/pyk/pull/835, the
function `build_rule` is modified to maintain all structure of a rule.
But that routine is also used for creating pretty-printing rules for
parsing via the frontend. Rules that need to be parsed via the frontend
must not contain the `<generated*>` cells.

This PR makes sure that `KCFGShow(...)` still minimizes the rules that
are printed via `to_module`. This should allow undoing most of the
output changes here:
https://github.com/runtimeverification/kontrol/pull/325.

---------

Co-authored-by: devops <devops@runtimeverification.com>